Перенесен метод открытия журнала списания километража
